pyproject.toml
[project]
name = "prefix"
version = "0.1.0"
description = "Prefix-aware label smoothing experiments for Llama-3–style LM training"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
  "datasets>=2.19",
  "huggingface-hub>=0.36.0",
  "lm-eval>=0.4.9.2",
  "marisa-trie>=1.3.1",
  "matplotlib>=3.8",
  "mosaicml-streaming>=0.9",
  "pandas>=2.2",
  "pyyaml>=6.0",
  "seaborn>=0.13",
  "torch>=2.2",
  "transformers>=4.40",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
fix = true
show-fixes = true
